First version of working program
Program is fully working. Form can be build by adding inputs/subinputs
and by insterting specific criteria.

Form is being build recursively - component with subinputs calls itself
cheching if it has any subinputs, if so - it renders it and so on. The
same principle is taken when the form result is displayed.

Every data is stored on indexedDB, even after shuting down the tab
result will be remembered.